[Plugins] Add CImgPlugin: using CImg, add new loader for images
If loaded in a scene ("RequiredPlugin"), the user will be able
to load png,jpeg,bmp and tiff image files.
This plugin is supposed to replace the ImageQt class, as
it is imposing us to have Qt to load jpeg image (so jpeg was only
loadable if a Qt GUI was activated). ImagePNG and ImageBMP could
be replaced too (to be discussed)
CImg needs libpng (PNG), libjpeg (JPEG), libtiff (TIFF).
Those formats will be disabled if those libs are not found at
the CMake step.
TODO: save image according to extension, add imagemagick lib for more
exotic formats.